refactor(Parser, Validator): Move validation methods from Parser to Validator class
