+++ src/main/java/calculator/util/Validator.java
@@ -2,7 +2,10 @@ package calculator.util;
 
 import static calculator.message.Delimiter.CUSTOM_DELIMITER;
 import static calculator.message.ErrorMessage.CUSTOM_DELIMITER_CANNOT_BE_NUMBER_MESSAGE;
+import static calculator.message.ErrorMessage.EMPTY_ELEMENT_MESSAGE;
 import static calculator.message.ErrorMessage.EMPTY_INPUT_MESSAGE;
+import static calculator.message.ErrorMessage.INVALID_ELEMENT_MESSAGE;
+import static calculator.message.ErrorMessage.INVALID_FLOAT_INPUT_MESSAGE;
 import static calculator.message.ErrorMessage.NEGATIVE_NUMBER_MESSAGE;
 
 import java.util.List;
@@ -13,12 +16,6 @@ public class Validator {
         String newInput = validateCustomDelimiter(userInput);
 
         return newInput;
-    }
-
-    private static void isBlank(String str) {
-        if (str.isBlank()) {
-            throw new IllegalArgumentException(EMPTY_INPUT_MESSAGE);
-        }
     }
 
     public static void isNegativeNumber(List<Integer> numbers) {
@@ -37,6 +34,48 @@ public class Validator {
         }
 
         return str;
+    }
+
+    public static void validateElement(String element) {
+        if (isBlankBetweenDelimiters(element)) {
+            throw new IllegalArgumentException(EMPTY_ELEMENT_MESSAGE);
+        }
+
+        if (isFloat(element)) {
+            throw new IllegalArgumentException(INVALID_FLOAT_INPUT_MESSAGE);
+        }
+
+        if (!isInteger(element)) {
+            throw new IllegalArgumentException(INVALID_ELEMENT_MESSAGE);
+        }
+    }
+
+    private static void isBlank(String str) {
+        if (str.isBlank()) {
+            throw new IllegalArgumentException(EMPTY_INPUT_MESSAGE);
+        }
+    }
+
+    private static boolean isBlankBetweenDelimiters(String element) {
+        return element.isBlank();
+    }
+
+    private static boolean isFloat(String element) {
+        try {
+            Float.parseFloat(element);
+            return element.contains(".");
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
+
+    public static boolean isInteger(String element) {
+        try {
+            Integer.parseInt(element);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
     }
 
     private static void validateCustomDelimiterIsNotNumber() {
